ep: Move "pure" diffing functions to zeta_prompt:udiff
We need this so that zeta_prompt could do expected_output formatting for
training. Keeping everything prompt-related in a pure module makes it
easier to write bindings.
